feat: Money 클래스가 int형 생성자에서 받을 수 있도록 수정

+++ src/main/java/lotto/domain/WinningInformation.java
@@ -1,19 +1,18 @@
 package lotto.domain;
 
-import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public enum WinningInformation {
 
-    NOT_MATCHES(0, new Money(BigInteger.valueOf(0))),
-    ONE_MATCHES(1, new Money(BigInteger.valueOf(0))),
-    TWO_MATCHES(2, new Money(BigInteger.valueOf(0))),
-    THREE_MATCHES(3, new Money(BigInteger.valueOf(5_000))),
-    FOUR_MATCHES(4, new Money(BigInteger.valueOf(50_000))),
-    FIVE_MATCHES(5, new Money(BigInteger.valueOf(1_500_000))),
-    SIX_MATCHES(6, new Money(BigInteger.valueOf(2_000_000_000)));
+    NOT_MATCHES(0, new Money(0)),
+    ONE_MATCHES(1, new Money(0)),
+    TWO_MATCHES(2, new Money(0)),
+    THREE_MATCHES(3, new Money(5_000)),
+    FOUR_MATCHES(4, new Money(50_000)),
+    FIVE_MATCHES(5, new Money(1_500_000)),
+    SIX_MATCHES(6, new Money(2_000_000_000));
 
     private final int matchesCount;
     private final Money amount;
