+++ src/test/java/lotto/domain/WinningInformationTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -46,6 +45,6 @@ class WinningInformationTest {
 
         Money actual = WinningInformation.sumAmounts(winningInformations);
 
-        assertThat(actual).isEqualTo(new Money(BigInteger.valueOf(2_001_555_000)));
+        assertThat(actual).isEqualTo(new Money(2_001_555_000));
     }
 }
