+++ src/test/java/lotto/domain/WinningStatisticsTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 
@@ -53,7 +52,7 @@ class WinningStatisticsTest {
     @DisplayName("수익률을 반환한다.")
     @Test
     void calculateYield() {
-        BigDecimal actual = winningStatistics.calculateYield(new Money(BigInteger.valueOf(14000)));
+        BigDecimal actual = winningStatistics.calculateYield(new Money(14000));
 
         assertThat(actual).isEqualTo(BigDecimal.valueOf(285714.64));
     }
